Fix epoch issue and renamed packages.
